Rename templates action types file
