+++ src/store/templates/templates.actionTypes.js
